Added Plane. New scene. Improvements in tracer

+++ lab2/src/iLight.h
@@ -12,5 +12,5 @@ public:
    iLight(const Eigen::Vector3f &c, const Eigen::Vector3f &p): color(c), pos(p) {};
    virtual ~iLight() {};
 
-   virtual double Shadow(const cScene &scene, const Eigen::Vector3f &p, Eigen::Vector3f &l) const = 0;
+   virtual float Shadow(const cScene &scene, const Eigen::Vector3f &p, Eigen::Vector3f &l) const = 0;
 };
